pyproject.toml
[project]
name = "domain-adaptation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "pillow>=12.0.0",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "tqdm>=4.67.1",
]
